+++ get_ip.py
@@ -1,0 +1,5 @@
+
+import websockets
+import socket
+myip = [(s.connect(('8.8.8.8', 53)), s.getsockname()[0], s.close()) for s in [socket.socket(socket.AF_INET, socket.SOCK_DGRAM)]][0][1]
+
